Progressively enhanced form actions
This enables forms to be submitted even before hydration.

While handling an MPA-style POST request, we can parse the hidden form
fields that contain meta data about the form action, using
`decodeAction` and `decodeFormState`, to call the action, and send the
result as form state, along with the rendered root, in the RSC response.
The form state is used during server-side rendering (passed into
`renderToReadableStream`) to render the result as part of the initial
HTML, as well as in the client during hydration (passed into
`hydrateRoot`) to avoid hydration mismatches.

BREAKING CHANGE: The Webpack RSC server plugin now emits the server
manifest in the structure that React expects.