Pass `formState` directly into `createHtmlStream`
This avoids the hacky implementation of assigning the form state after server-side rendering has already begun. I believe in the React flight fixture this was only done because RSC rendering and SSR happens in two separate servers. So the form state needed to be extracted from the RSC response stream. Here, we can just pass it around. It seems Next.js is doing that as well (in a more convoluted way, so it's a bit complicated to follow the flow there): https://github.com/vercel/next.js/blob/canary/packages/next/src/server/app-render/app-render.tsx

BREAKING CHANGE: The `fomState` now needs to be passed in as part of the options of `createHtmlStream`.