Avoid duplicate work being done in RSC server plugin
The RSC server loader already collects all the server references. So we
don't need to do the same work again in the plugin. Instead, we can just
read them from the provided map.

This should simplify adding support for inline `'use server'` references
to the plugin.

BREAKING CHANGE: The `serverReferencesMap` is now a required option of
`createWebpackRscServerLoader`. In addition, it also needs to be passed
to `createWebpackRscSsrLoader` (required for server actions that are
imported from client components).