Move `hydrateApp` to `@mfng/core/client/browser`
This is imported from the client/browser entry, so it should be in
`@mfng/core/client/browser`, and not in `@mfng/core/client`. The latter
has modules that can be used in both, the SSR and browser client. Since
`Router` is also imported through the same barrel file, we can remove
the `'use client'` directive from it.

Continuation from #53 and #54