Fix `TypeError [ERR_INVALID_STATE]: Invalid state: Unable to enqueue`
This error occurred when a request was cancelled by the user. This is
logged by React as:

```
The render was aborted by the server without a reason.
```

In this case the HTML stream was already closed, and the buffering
transform stream could not enqueue any more chunks.

Fortunately, after a bit of testing, it turned out that the whole
buffering is not needed anymore. I also couldn't find similar code like
this in Next.js anymore. So we are just removing it as well.

The injection of the RSC chunks, and the hydration still work fine
without it.